Moved the code to separate files
